kcd's version of extra 1
